refactor(ai-service): split transformer model — extract scaler + dataset
Third and largest Page-5 split. Drops transformer/model.py from 2737
to 1890 lines (-847 LOC, -31%) by pulling out the two parts that
weren't model code: per-feature scaling and the streaming dataset.

Extracted:
  app/ml/transformer/model.py
    ├─ PerFeatureScaler                    ┐
    │   (264 lines — sklearn wrappers,     ├─►  app/storage/datasets/transformer_scaler.py
    │    no torch, no I/O beyond joblib)   │
    ├─ _RunningFeatureStats (24 lines) ────┘
    │
    └─ TelemetryActionDataset (558 lines) ──►  app/storage/datasets/telemetry_dataset.py

Both new files live in app/storage/datasets/ rather than ml/, which
was a one-step deviation from the diagram's Page-5 placement
(ml/transformer/scaler.py). Reason: when scaler.py first landed at
ml/transformer/, import-linter flagged it — the dataset (storage)
was reaching up into ml to grab the scaler. Architecturally the
scaler IS data-prep (numpy + sklearn, no torch); it's used upstream
of the model. Moved it next to the dataset that consumes it, where
both can now be imported from the same band without violating
storage-no-upward.

The trainer (ExpertActionTrainer, 784 lines) and the async entry
function (prepare_and_train_coach_transformer_model) stay in model.py
for now — moving them into pipelines/training/ would require unwiring
a lot of method-level state references and is its own design pass.

Validated:
  • Identity preserved through both new and old paths
    (`from app.ml.transformer.model import PerFeatureScaler` and
    `from app.storage.datasets.transformer_scaler import PerFeatureScaler`
    return the same object)
  • PerFeatureScaler + _RunningFeatureStats construct cleanly
  • lint-imports 11 kept, 0 broken
  • pytest 24 passed
  • /health 200 after container restart
